refactor: all listeners are moved to individual functions
